feat: Me 수정 api 추가 ( 성별, 생년월일, 키, 몸무게, 예방 질병 )

+++ lib/data/models/me/ResponseMeInfoProfileModel.dart
@@ -6,8 +6,8 @@ part 'ResponseMeInfoProfileModel.g.dart';
 class ResponseMeInfoProfileModel {
   final String? gender;
   final String? birthday;
-  final double? height;
-  final double? weight;
+  final int? height;
+  final int? weight;
   final List<String>? diseases;
 
   ResponseMeInfoProfileModel({
